pr-context-mcp/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
cachetools==5.3.2
boto3==1.34.0
pydantic==2.5.0
